Faster state backend: propagate the errors from aggregators and reducers as values instead of panicking.
The errors unfortunately won't be returned by the .aggregate or .reduce methods on the active handles. Instead the value persisted in the backend will become effectively poisoned, returning an error on every .get(), until the state gets cleared. The returned error is not the same error that has been raised during the rmw, but its Display form instead (needed because of the serialization requirements enforced by faster-rs).