Add ArconUnit type
prost does not support the unit type `()` in messages. The closest thing
to having a unit type in prost is to add an empty sub-message. This commit
introduces the ArconUnit type which can be used in-place of `()`.

For example, this does not work:

    #[prost::Message]
    struct Event {
        x: ()
    }

but this works:

    #[prost::Message]
    struct Event {
        #[prost(message, required)]
        x: arcon::prelude::ArconUnit
    }

this becomes more important in enums, this does not work:

    #[prost::Oneof]
    enum List {
        #[prost(message, required)]
        Cons(Cons),
        Nil
    }

one may have:

    #[prost::Oneof]
    enum List {
        #[prost(message, required)]
        Cons(Cons),
        #[prost(message, required)]
        Nil(arcon::prelude::ArconUnit)
    }