Add macro for deriving prost boilerplate
This macro makes it possible to derive prost without specifying any
tags using an #[arcon::proto] macro.

    #[arcon::proto]
    struct Event {
        x: i32,
        y: f32,
    }

which expands to:

    #[prost::Message]
    struct Event {
        #[prost(int32)]
        x: i32,
        #[prost(float)]
        y: f32,
    }

It relies on parsing types, so it assumes that the types `i32` and `f32`
are not aliased for something else. If they are, then prost will throw
an error in either case.

See arcon_tests/src/proto_derive_test.rs for more usage examples that
cover enums also.