Add macro for deriving prost boilerplate (#172)
* Add ArconUnit type

prost does not support the unit type `()` in messages. The closest thing
to having a unit type in prost is to add an empty sub-message. This commit
introduces the ArconUnit type which can be used in-place of `()`.

For example, this does not work:

    #[prost::Message]
    struct Event {
        x: ()
    }

but this works:

    #[prost::Message]
    struct Event {
        #[prost(message, required)]
        x: arcon::prelude::ArconUnit
    }

this becomes more important in enums, this does not work:

    #[prost::Oneof]
    enum List {
        #[prost(message, required)]
        Cons(Cons),
        Nil
    }

one may have:

    #[prost::Oneof]
    enum List {
        #[prost(message, required)]
        Cons(Cons),
        #[prost(message, required)]
        Nil(arcon::prelude::ArconUnit)
    }

* Add macro for deriving prost boilerplate

This macro makes it possible to derive prost without specifying any
tags using an #[arcon::proto] macro.

    #[arcon::proto]
    struct Event {
        x: i32,
        y: f32,
    }

which expands to:

    #[prost::Message]
    struct Event {
        #[prost(int32)]
        x: i32,
        #[prost(float)]
        y: f32,
    }

It relies on parsing types, so it assumes that the types `i32` and `f32`
are not aliased for something else. If they are, then prost will throw
an error in either case.

See arcon_tests/src/proto_derive_test.rs for more usage examples that
cover enums also.